OpenJDK20: finish NetBSD support
The copyright year, source date was added to get around issues with NetBSD's
stock date program.  However, the build failed again later without any kind
of workaround.  The solution was the replace date in the system root with
the coreutils version.

I guess it means theoretically a user could run into the same problem with
NetBSD's date.  the solution in this case would be to hardcode date to
run the coreutils version and make coreutils a run dependency.